requirements.txt
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-phonenumber-field==1.3.0
djangorestframework==3.7.3
httpie==0.9.9
idna==2.6
phonenumberslite==8.8.8
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
stripe==1.75.3
urllib3==1.22
